declare new authentication related errors
